run tests for both CJS and ESM versions
