+++ tests/test.cjs
@@ -1,0 +1,4 @@
+let pc = require("../picocolors.cjs");
+let { runTestSuite } = require("./suite.cjs");
+
+runTestSuite(pc);
