+++ tests/test.js
@@ -1,0 +1,4 @@
+import * as pc from "../picocolors.js";
+import { runTestSuite } from "./suite.cjs";
+
+runTestSuite(pc);
